Tender, binding: Fix build failures
