+++ bindings/muen/muen-net.c
@@ -31,12 +31,13 @@ static uint8_t mac_addr[6];
 
 solo5_result_t solo5_net_write(const uint8_t *buf, size_t size)
 {
-    return shm_net_write(net_out, buf, size);
+    return shm_to_solo5_result(shm_net_write(net_out, buf, size));
 }
 
 solo5_result_t solo5_net_read(uint8_t *buf, size_t size, size_t *read_size)
 {
-    return shm_net_read(net_in, &net_rdr, buf, size, read_size);
+    return shm_to_solo5_result(shm_net_read(net_in, &net_rdr, buf,
+                size, read_size));
 }
 
 bool muen_net_pending_data()
